+++ src/main/java/com/github/gtexpert/gtwp/module/Modules.java
@@ -38,6 +38,7 @@ public class Modules implements IModuleContainer {
     public static final String MODULE_RTHINGS = "rthings_integration";
     public static final String MODULE_RUSTIC = "rustic_integration";
     public static final String MODULE_IDS = "ids_integration";
+    public static final String MODULE_MWORLD = "mworld_integration";
 
     @Override
     public String getID() {