+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 - Extra Utilities 2
 - Integrated Dynamics
 - Random Things
+- Rustic
+- Mystical World
+- Roots
 
 * * *